requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.9.0
yfinance
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
websocket-client